User commit

+++ src/main/java/com/example/demo/repository/TimeStampRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.TimeStampEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimeStampRepository extends JpaRepository<TimeStampEntity, Integer> {
+public interface TimeStampRepository extends JpaRepository<TimeStampEntity, Long> {
 }
